Add files via upload
Open the esp123.ino file in the Arduino IDE, compile the code, and successfully upload it to the ESP32 microcontroller. Create a dedicated folder named temple and store the index.html file within it. Additionally, set up a separate folder named __pycache__ and place sensor_data.cpython-313.pyc inside it. Ensure that the exchangeable wire for the smart safety helmet is properly connected before proceeding.